CREATE errors.py
